backend/requirements.txt
# LogiSight backend — Phase 1 core
fastapi>=0.110.0
python-multipart>=0.0.9
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Phase 2 — Supabase JWT (RS256 / JWKS)
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
supabase>=2.0.0

# Phase 4 — Copilot (LangChain SQL Agent)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
psycopg2-binary>=2.9.9

# Invoice Extraction
veryfi>=3.0.0
